+++ src/lib.rs
@@ -1,26 +1,25 @@
 #![deny(missing_docs)]
 
-//! This crate provides a [egui](https://github.com/emilk/egui) integration for the [Bevy](https://github.com/bevyengine/bevy) game engine.
+//! This crate provides a [Egui](https://github.com/emilk/egui) integration for the [Bevy](https://github.com/bevyengine/bevy) game engine.
+//!
+//! **Trying out:**
+//!
+//! An example WASM project is live at [mvlabat.github.io/bevy_egui_web_showcase](https://mvlabat.github.io/bevy_egui_web_showcase/index.html) [[source](https://github.com/mvlabat/bevy_egui_web_showcase)].
 //!
 //! **Features:**
 //! - Desktop and web ([bevy_webgl2](https://github.com/mrk-its/bevy_webgl2)) platforms support
 //! - Clipboard (web support is limited to the same window, see [rust-windowing/winit#1829](https://github.com/rust-windowing/winit/issues/1829))
 //! - Opening URLs
+//! - Multiple windows support (see [./examples/two_windows.rs](./examples/two_windows.rs))
 //!
 //! `bevy_egui` can be compiled with using only `bevy` and `egui` as dependencies: `manage_clipboard` and `open_url` features,
 //! that require additional crates, can be disabled.
-//!
-//! ## Trying out
-//!
-//! An example WASM project is live at [mvlabat.github.io/bevy_egui_web_showcase](https://mvlabat.github.io/bevy_egui_web_showcase/index.html) [[source](https://github.com/mvlabat/bevy_egui_web_showcase)].
-//!
-//! **Note** that in order to use `bevy_egui`in WASM you need [bevy_webgl2](https://github.com/mrk-its/bevy_webgl2) of at least `0.5.0` version.
 //!
 //! ## Usage
 //!
 //! Here's a minimal usage example:
 //!
-//! ```no_run
+//! ```no_run,rust
 //! use bevy::prelude::*;
 //! use bevy_egui::{egui, EguiContext, EguiPlugin};
 //!
@@ -32,9 +31,8 @@
 //!         .run();
 //! }
 //!
-//! fn ui_example(mut egui_context: ResMut<EguiContext>) {
-//!     let ctx = &mut egui_context.ctx;
-//!     egui::Window::new("Hello").show(ctx, |ui| {
+//! fn ui_example(egui_context: Res<EguiContext>) {
+//!     egui::Window::new("Hello").show(egui_context.ctx(), |ui| {
 //!         ui.label("world");
 //!     });
 //! }
@@ -43,11 +41,12 @@
 //! For a more advanced example, see [examples/ui.rs](examples/ui.rs).
 //!
 //! ```bash
-//! cargo run --example ui --features="bevy/x11 bevy/png bevy/bevy_wgpu"
+//! cargo run --example ui
 //! ```
 //!
 //! ## See also
 //!
+//! - [`bevy-inspector-egui`](https://github.com/jakobhellermann/bevy-inspector-egui)
 //! - [`bevy_megaui`](https://github.com/mvlabat/bevy_megaui)
 
 pub use egui;
@@ -60,9 +59,12 @@ use crate::{egui_node::EguiNode, systems::*, transform_node::EguiTransformNode};
 use bevy::{
     app::{AppBuilder, CoreStage, Plugin},
     asset::{Assets, Handle, HandleUntyped},
+    ecs::{
+        schedule::{ParallelSystemDescriptorCoercion, StageLabel, SystemLabel, SystemStage},
+        system::IntoSystem,
+    },
     input::InputSystem,
     log,
-    prelude::{IntoSystem, ParallelSystemDescriptorCoercion, StageLabel, SystemLabel, SystemStage},
     reflect::TypeUuid,
     render::{
         pipeline::{
@@ -75,16 +77,17 @@ use bevy::{
         texture::{Texture, TextureFormat},
         RenderStage,
     },
+    utils::HashMap,
+    window::WindowId,
 };
 #[cfg(all(feature = "manage_clipboard", not(target_arch = "wasm32")))]
 use clipboard::{ClipboardContext, ClipboardProvider};
 #[cfg(all(feature = "manage_clipboard", not(target_arch = "wasm32")))]
 use std::cell::{RefCell, RefMut};
-use std::collections::HashMap;
 #[cfg(all(feature = "manage_clipboard", not(target_arch = "wasm32")))]
 use thread_local::ThreadLocal;
 
-/// A handle pointing to the egui [PipelineDescriptor].
+/// A handle pointing to the egui [`PipelineDescriptor`].
 pub const EGUI_PIPELINE_HANDLE: HandleUntyped =
     HandleUntyped::weak_from_u64(PipelineDescriptor::TYPE_UUID, 9404026720151354217);
 /// Name of the transform uniform.
@@ -92,10 +95,10 @@ pub const EGUI_TRANSFORM_RESOURCE_BINDING_NAME: &str = "EguiTransform";
 /// Name of the texture uniform.
 pub const EGUI_TEXTURE_RESOURCE_BINDING_NAME: &str = "EguiTexture_texture";
 
-/// Adds all egui resources and render graph nodes.
+/// Adds all Egui resources and render graph nodes.
 pub struct EguiPlugin;
 
-/// A resource containing global UI settings.
+/// A resource for storing global UI settings.
 #[derive(Clone, Debug, PartialEq)]
 pub struct EguiSettings {
     /// Global scale factor for egui widgets (`1.0` by default).
@@ -120,8 +123,9 @@ impl Default for EguiSettings {
     }
 }
 
-/// A resource that stores the input passed to Egui.
-/// It gets reset during the [EguiSystem::ProcessInput] system.
+/// Is used for storing the input passed to Egui. The actual resource is a [`HashMap<WindowId, EguiInput>`].
+///
+/// It gets reset during the [`EguiSystem::ProcessInput`] system.
 #[derive(Clone, Debug, Default)]
 pub struct EguiInput {
     /// Egui's raw input.
@@ -129,7 +133,8 @@ pub struct EguiInput {
 }
 
 /// A resource for accessing clipboard.
-/// Is available only if `manage_clipboard` feature is enabled.
+///
+/// The resource is available only if `manage_clipboard` feature is enabled.
 #[cfg(feature = "manage_clipboard")]
 #[derive(Default)]
 pub struct EguiClipboard {
@@ -146,7 +151,7 @@ impl EguiClipboard {
         self.set_contents_impl(contents);
     }
 
-    /// Gets clipboard contents. Returns [None] if clipboard provider is unavailable or returns an error.
+    /// Gets clipboard contents. Returns [`None`] if clipboard provider is unavailable or returns an error.
     pub fn get_contents(&self) -> Option<String> {
         self.get_contents_impl()
     }
@@ -198,26 +203,25 @@ impl EguiClipboard {
     }
 }
 
+/// Is used for storing Egui shapes. The actual resource is [`HashMap<WindowId, EguiShapes>`].
 #[derive(Clone, Default)]
-/// A resource for storing Egui shapes.
 pub struct EguiShapes {
     /// Pairs of rectangles and paint commands.
     ///
-    /// The field gets populated during the [EguiStage::UiFrameEnd] stage and reset during `EguiNode::update`.
+    /// The field gets populated during the [`EguiStage::UiFrameEnd`] stage and reset during `EguiNode::update`.
     pub shapes: Vec<egui::paint::ClippedShape>,
 }
 
-/// A resource for storing Egui output.
+/// Is used for storing Egui output. The actual resource is [`HashMap<WindowId, EguiOutput>`].
 #[derive(Clone, Default)]
 pub struct EguiOutput {
-    /// The field gets updated during the [EguiStage::UiFrameEnd] stage.
+    /// The field gets updated during the [`EguiStage::UiFrameEnd`] stage.
     pub output: egui::Output,
 }
 
-/// A resource that is used to store `bevy_egui` context.
+/// A resource for storing `bevy_egui` context.
 pub struct EguiContext {
-    /// Egui context.
-    pub ctx: egui::CtxRef,
+    ctx: HashMap<WindowId, egui::CtxRef>,
     egui_textures: HashMap<egui::TextureId, Handle<Texture>>,
     mouse_position: Option<(f32, f32)>,
 }
@@ -225,10 +229,28 @@ pub struct EguiContext {
 impl EguiContext {
     fn new() -> Self {
         Self {
-            ctx: Default::default(),
+            ctx: HashMap::default(),
             egui_textures: Default::default(),
             mouse_position: Some((0.0, 0.0)),
         }
+    }
+
+    /// Egui context of the primary window.
+    #[track_caller]
+    pub fn ctx(&self) -> &egui::CtxRef {
+        &self.ctx[&WindowId::primary()]
+    }
+
+    /// Egui context for a specific window.
+    /// If you want to display UI on a non-primary window,
+    /// make sure to set up the render graph by calling [`setup_pipeline`].
+    #[track_caller]
+    pub fn ctx_for_window(&self, window: WindowId) -> &egui::CtxRef {
+        &self
+            .ctx
+            .get(&window)
+            .ok_or_else(|| format!("window with id {} not found", window))
+            .unwrap()
     }
 
     /// Can accept either a strong or a weak handle.
@@ -301,16 +323,17 @@ pub mod node {
 #[derive(StageLabel, Clone, Hash, Debug, Eq, PartialEq)]
 /// The names of `bevy_egui` stages.
 pub enum EguiStage {
-    /// Runs before [bevy::render::RenderStage::RenderResource]. This is where we read Egui's output.
+    /// Runs before [`bevy::render::RenderStage::RenderResource`]. This is where we read Egui's output.
     UiFrameEnd,
 }
 
 #[derive(SystemLabel, Clone, Hash, Debug, Eq, PartialEq)]
 /// The names of egui systems.
 pub enum EguiSystem {
-    /// Reads keyboard, mouse etc. input and write it into the [`EguiInput`] resource
+    /// Reads Egui inputs (keyboard, mouse, etc) and writes them into the [`EguiInput`] resource.
     ///
-    /// To modify the input, specify
+    /// To modify the input, you can hook your system like this:
+    ///
     /// `system.after(EguiSystem::ProcessInput).before(EguiSystem::BeginFrame)`.
     ProcessInput,
     /// Begins the `egui` frame
@@ -348,13 +371,13 @@ impl Plugin for EguiPlugin {
 
         let world = app.world_mut();
         world.get_resource_or_insert_with(EguiSettings::default);
-        world.get_resource_or_insert_with(EguiInput::default);
-        world.get_resource_or_insert_with(EguiOutput::default);
-        world.get_resource_or_insert_with(EguiShapes::default);
+        world.get_resource_or_insert_with(HashMap::<WindowId, EguiInput>::default);
+        world.get_resource_or_insert_with(HashMap::<WindowId, EguiOutput>::default);
+        world.get_resource_or_insert_with(HashMap::<WindowId, WindowSize>::default);
+        world.get_resource_or_insert_with(HashMap::<WindowId, EguiShapes>::default);
         #[cfg(feature = "manage_clipboard")]
         world.get_resource_or_insert_with(EguiClipboard::default);
         world.insert_resource(EguiContext::new());
-        world.insert_resource(WindowSize::new(0.0, 0.0, 0.0));
 
         let world = world.cell();
 
@@ -368,58 +391,89 @@ impl Plugin for EguiPlugin {
             EGUI_PIPELINE_HANDLE,
             build_egui_pipeline(&mut shaders, msaa.samples),
         );
+
         let mut render_graph = world.get_resource_mut::<RenderGraph>().unwrap();
+        setup_pipeline(&mut render_graph, &msaa, RenderGraphConfig::default());
+    }
+}
 
-        render_graph.add_node(node::EGUI_PASS, EguiNode::new(&msaa));
-        render_graph
-            .add_node_edge(base::node::MAIN_PASS, node::EGUI_PASS)
-            .unwrap();
+/// Egui's render graph config.
+#[allow(missing_docs)]
+pub struct RenderGraphConfig {
+    pub window_id: WindowId,
+    pub egui_pass: &'static str,
+    pub main_pass: &'static str,
+    pub swap_chain_node: &'static str,
+    pub depth_texture: &'static str,
+    pub sampled_color_attachment: &'static str,
+    pub transform_node: &'static str,
+}
 
-        if let Ok(ui_pass) = render_graph.get_node_id(bevy::ui::node::UI_PASS) {
-            render_graph
-                .add_node_edge(ui_pass, node::EGUI_PASS)
-                .unwrap();
+impl Default for RenderGraphConfig {
+    fn default() -> Self {
+        RenderGraphConfig {
+            window_id: WindowId::primary(),
+            egui_pass: node::EGUI_PASS,
+            main_pass: base::node::MAIN_PASS,
+            swap_chain_node: base::node::PRIMARY_SWAP_CHAIN,
+            depth_texture: base::node::MAIN_DEPTH_TEXTURE,
+            sampled_color_attachment: base::node::MAIN_SAMPLED_COLOR_ATTACHMENT,
+            transform_node: node::EGUI_TRANSFORM,
         }
+    }
+}
 
+/// Set up egui render pipeline.
+///
+/// The pipeline for the primary window will already be set up by the [`EguiPlugin`],
+/// so you'll only need to manually call this if you want to use multiple windows.
+pub fn setup_pipeline(render_graph: &mut RenderGraph, msaa: &Msaa, config: RenderGraphConfig) {
+    render_graph.add_node(config.egui_pass, EguiNode::new(&msaa, config.window_id));
+    render_graph
+        .add_node_edge(config.main_pass, config.egui_pass)
+        .unwrap();
+    if let Ok(ui_pass) = render_graph.get_node_id(bevy::ui::node::UI_PASS) {
         render_graph
-            .add_slot_edge(
-                base::node::PRIMARY_SWAP_CHAIN,
-                WindowSwapChainNode::OUT_TEXTURE,
-                node::EGUI_PASS,
-                if msaa.samples > 1 {
-                    "color_resolve_target"
-                } else {
-                    "color_attachment"
-                },
-            )
-            .unwrap();
-
-        render_graph
-            .add_slot_edge(
-                base::node::MAIN_DEPTH_TEXTURE,
-                WindowTextureNode::OUT_TEXTURE,
-                node::EGUI_PASS,
-                "depth",
-            )
-            .unwrap();
-
-        if msaa.samples > 1 {
-            render_graph
-                .add_slot_edge(
-                    base::node::MAIN_SAMPLED_COLOR_ATTACHMENT,
-                    WindowSwapChainNode::OUT_TEXTURE,
-                    node::EGUI_PASS,
-                    "color_attachment",
-                )
-                .unwrap();
-        }
-
-        // Transform.
-        render_graph.add_system_node(node::EGUI_TRANSFORM, EguiTransformNode::new());
-        render_graph
-            .add_node_edge(node::EGUI_TRANSFORM, node::EGUI_PASS)
+            .add_node_edge(ui_pass, config.egui_pass)
             .unwrap();
     }
+    render_graph
+        .add_slot_edge(
+            config.swap_chain_node,
+            WindowSwapChainNode::OUT_TEXTURE,
+            config.egui_pass,
+            if msaa.samples > 1 {
+                "color_resolve_target"
+            } else {
+                "color_attachment"
+            },
+        )
+        .unwrap();
+    render_graph
+        .add_slot_edge(
+            config.depth_texture,
+            WindowTextureNode::OUT_TEXTURE,
+            config.egui_pass,
+            "depth",
+        )
+        .unwrap();
+    if msaa.samples > 1 {
+        render_graph
+            .add_slot_edge(
+                config.sampled_color_attachment,
+                WindowSwapChainNode::OUT_TEXTURE,
+                config.egui_pass,
+                "color_attachment",
+            )
+            .unwrap();
+    }
+    render_graph.add_system_node(
+        config.transform_node,
+        EguiTransformNode::new(config.window_id),
+    );
+    render_graph
+        .add_node_edge(config.transform_node, config.egui_pass)
+        .unwrap();
 }
 
 fn build_egui_pipeline(shaders: &mut Assets<Shader>, sample_count: u32) -> PipelineDescriptor {